api/requirements.txt
langchain
langchain_community
langchain_openai
langchain_xai
pydantic-ai
redis
uvicorn
fastapi
SQLAlchemy
alembic
python-dotenv
pydantic
psycopg2
passlib
PyJWT
celery
python-multipart
google-cloud-storage
pytest
httpx
python-dotenv